# documention/swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/national/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Auth"]
      }
    },
    "/national/auth/login/refresh": {
      "post": {
        "operationId": "AuthController_refreshLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshLoginDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Auth"]
      }
    },
    "/national/auth/forgotPassword": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ForgotPasswordDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Auth"]
      }
    },
    "/national/auth/resetPassword": {
      "put": {
        "operationId": "AuthController_resetPassword",
        "parameters": [
          {
            "name": "requestId",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PasswordResetDto" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["Auth"]
      }
    },
    "/national/auth/checkResetRequestId": {
      "put": {
        "operationId": "AuthController_checkResetRequestId",
        "parameters": [
          {
            "name": "requestId",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Auth"]
      }
    },
...
    "/national/analytics/getProjectsData": {
      "post": {
        "operationId": "AnalyticsController_getProjectsData",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProjectDataRequestDTO" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Analytics"]
      }
    },
...
    "/national/analytics/getProjectStatusSummary": {
      "post": {
        "operationId": "AnalyticsController_getProjectStatusSummary",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProjectDataRequestDTO" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Analytics"]
      }
    },
    "/national/analytics/getProjectsByStatusDetail": {
      "post": {
        "operationId": "AnalyticsController_getProjectsByStatusDetail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProjectDataRequestDTO" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Analytics"]
      }
    },
    "/national/analytics/getProjectCountBySector": {
      "post": {
        "operationId": "AnalyticsController_getProjectCountBySector",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProjectDataRequestDTO" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Analytics"]
      }
    },
    "/national/analytics/getProjectCountBySectorScope": {
      "post": {
        "operationId": "AnalyticsController_getProjectCountBySectorScope",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProjectDataRequestDTO" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Analytics"]
      }
    },
    "/national/analytics/getCreditSummary": {
      "post": {
        "operationId": "AnalyticsController_getCreditSummary",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProjectDataRequestDTO" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Analytics"]
      }
    },
    "/national/analytics/creditsSummaryByDate": {
      "post": {
        "operationId": "AnalyticsController_creditsSummaryByDate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProjectDataRequestDTO" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Analytics"]
      }
    },
...
  },
  "info": {
    "title": "National API",
    "description": "RESTful Web API Documentation.",
    "version": "0.5",
    "contact": {}
  },
...
  "servers": [{ "url": "/" }],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" },
      "api_key": { "type": "apiKey", "in": "header", "name": "api_key" }
    },
    "schemas": {
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" }
        },
        "required": ["username", "password"]
      },
      "RefreshLoginDto": {
        "type": "object",
        "properties": { "refreshToken": { "type": "string" } },
        "required": ["refreshToken"]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": { "email": { "type": "string" } },
        "required": ["email"]
      },
      "PasswordResetDto": {
        "type": "object",
        "properties": { "newPassword": { "type": "string" } },
        "required": ["newPassword"]
      },
...
      "ProjectDataRequestDTO": {
        "type": "object",
        "properties": {
          "isMine": { "type": "boolean" },
          "timeZone": { "type": "number" },
          "startDate": { "type": "number" },
          "endDate": { "type": "number" },
          "sector": {
            "type": "string",
            "enum": [
              "Energy",
              "Agriculture",
              "Health",
              "Education",
              "Transport",
              "Manufacturing",
              "Hospitality",
              "Forestry",
              "Waste",
              "Other"
            ]
          }
        },
        "required": ["isMine", "timeZone", "startDate", "endDate", "sector"]
      },
...
    }
  }
}